requirements.txt
pycryptodome>=3.18.0
pillow>=9.0.0
tkinterdnd2 @ git+https://github.com/pmgagne/tkinterdnd2.git
